feat(export): if the svg graphviz generates contains &nbsp; make sure exports to png work, and exported svg's are valid again
also refactor(workarounds) - all workarounds for graphviz/ atom quirks are now in a workarounds.js module for easy maintenance/ removal
