+++ lib/graphviz-preview-plus-view.js
@@ -1,5 +1,4 @@
 "use babel";
-
 import {Emitter, Disposable, CompositeDisposable, File} from 'atom';
 import {$, $$$, ScrollView} from 'atom-space-pen-views';
 import path from 'path';
@@ -9,7 +8,7 @@ import fs from 'fs-plus';
 let renderGraphViz = null;
 let renderError = null;
 let svgToRaster = null;
-let resolveImage = null;
+let workarounds = null;
 
 let latestKnownEditorId = null;
 
@@ -320,68 +319,6 @@ export default class GraphVizPreviewView extends ScrollView {
         }
     }
 
-    resolveImages(pRenderedSVG, pWorkingDirectory) {
-        const lImages = pRenderedSVG.find('image');
-        if (Boolean(lImages) && resolveImage === null) {
-            resolveImage = require('./resolveImage');
-        }
-        for (let i = 0; i < lImages.length; i++) {
-            lImages[i].href.baseVal = resolveImage(lImages[i].href.baseVal, pWorkingDirectory);
-        }
-    }
-    /**
-     * GraphViz incorrectly renders SVG's in case there's a dpi !== the
-     * default (72) in the input dot. Root cause is that the viewBox doesn't
-     * get resized, while width, height and the main `g`'s scale are.
-     *
-     * This function corrects that by setting the viewBox to the width and
-     * height of the svg.
-     *
-     * Additional complexity: GraphViz uses pt as unit in the width & height.
-     * It uses those numbers 1:1 width & height in the viewBox. However, viewBox
-     * unit's are not points. So the SVG GraphViz outputs is always slightly
-     * bigger than it should be. graphviz-preview-plus does not interfere in
-     * that behavior.
-     *
-     * @param  {object} pRenderedSVG the (jQuery object wrapped) SVG as rendered
-     *                               by GraphViz
-     * @return {nothing}             This function _directly_ manipulates the
-     *                               input SVG
-     */
-    correctViewBox(pRenderedSVG) {
-        const lWidth = Number.parseFloat(pRenderedSVG.attr("width"));
-        const lHeight = Number.parseFloat(pRenderedSVG.attr("height"));
-
-        // should have been able to use pRenderedSVG.attr to set the viewBox,
-        // however that (jQuery) function executes a toLowerCase (or similar),
-        // As svg attributes are case sensitive this doesn't have any effect.
-        // Hence direct DOM manipulation:
-        const lRenderedSVGDOMElement = pRenderedSVG.get()[0];
-        lRenderedSVGDOMElement.setAttribute("viewBox", `0 0 ${lWidth} ${lHeight}`);
-    }
-
-    /**
-     * Usualy links in html in atom preview windows are clickable. However,
-     * correct links in svg (in the xlink namespace) aren't, whereas other
-     * browser implementations do.
-     *
-     * This is a workaround for that limitation. It adds a `href` attribute
-     * to each `a` in the svg in the preview window (_not_ to the exported svg!)
-     * that has an`xlink:href` attribute.
-     *
-     * @param  {object} pRenderedSVG the (jQuery object wrapped) SVG as rendered
-     *                               by GraphViz
-     * @return {nothing}             This function _directly_ manipulates the
-     *                               input SVG
-     */
-    makeLinksClickable(pRenderedSVG) {
-        pRenderedSVG.find("a").each(function() {
-            if (Boolean(this.getAttribute("xlink:href"))) {
-                this.setAttribute("href", this.getAttribute("xlink:href"));
-            }
-        });
-    }
-
     renderDotText(pText, pWorkingDirectory) {
         if (latestKnownEditorId !== this.editorId) {
             latestKnownEditorId = this.editorId;
@@ -396,20 +333,23 @@ export default class GraphVizPreviewView extends ScrollView {
             } else {
                 this.loading = false;
                 this.loaded = true;
-                this.svg = pSvg;
+                if (workarounds === null) {
+                    workarounds = require('./workarounds');
+                }
 
                 this.imageContainer.html(pSvg);
                 this.renderedSVG = this.imageContainer.find('svg');
-                /* start workarounds */
-                this.correctViewBox(this.renderedSVG);
-                this.makeLinksClickable(this.renderedSVG);
-                /* end workarounds */
-                this.svg = this.renderedSVG[0].outerHTML;
+                workarounds.correctViewBox(this.renderedSVG);
+                this.svg = workarounds.defineEntities(this.renderedSVG[0].outerHTML);
+
+                // this workaround only for the svg in the preview window
+                // ... not in the exported one
+                workarounds.makeLinksClickable(this.renderedSVG);
 
                 this.originalWidth = this.renderedSVG.attr('width');
                 this.originalHeight = this.renderedSVG.attr('height');
                 if (Boolean(pWorkingDirectory)) {
-                    this.resolveImages(this.renderedSVG, pWorkingDirectory);
+                    workarounds.resolveImages(this.renderedSVG, pWorkingDirectory);
                 }
 
                 if (this.mode === 'zoom-to-fit') {
