+++ lib/resolveImage.js
@@ -1,9 +1,8 @@
 "use babel";
-import * as path from 'path';
+import path from 'path';
 
 const imagePathIsAbsolute = (pImagePath) =>
     pImagePath.startsWith('/') || pImagePath.match(/^(https?|atom|file):\/\//);
-
 
 /**
  * resolves the given path to an image (in pImagePath) following these rules
